+++ rockpool/devices/dynapse/simulation/__init__.py
@@ -10,6 +10,19 @@ This module also provides a surrogate function implementation to support gradien
 See also:
     The neuron model tutorial provided in :ref:`/devices/DynapSE/neuron-model.ipynb`
 """
-from .dynapsim import *
-from .surrogate import *
-from .mismatch_prototype import *
+from rockpool.utilities.backend_management import (
+    backend_available,
+    missing_backend_shim,
+)
+
+if backend_available("jax"):
+    from .dynapsim import *
+    from .surrogate import *
+    from .mismatch_prototype import *
+else:
+    DynapSim = missing_backend_shim("DynapSim", "jax")
+    step_pwl = missing_backend_shim("step_pwl", "jax")
+    frozen_mismatch_prototype = missing_backend_shim("frozen_mismatch_prototype", "jax")
+    dynamic_mismatch_prototype = missing_backend_shim(
+        "dynamic_mismatch_prototype", "jax"
+    )